+++ CHANGES.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-0.10.2 (unreleased)
+0.10.2 (2017-06-29)
 -------------------
 
 - Add publication date check to display it on index view items : #17895